Allow for custom_component cloning from upstream
This patch allows the user to provide information about upstream custom components that should be
included in the Home Assistant instance. The helm chart will clone those repositories. Since not all
custom component repositories are structured such that the root directory is the directory to be put
into /config/custom_components this patch first clones the custom component repository into /tmp.
From there some additional processing may be required to extract the correct directory from the git
repository and to move it to /config/custom_components. Currently a componentPath is provided to indicate
the path to the custom component directory within the git repository. Currently, the helm chart does
a crude mv using this componentPath to populate the /custom_components directory.

Alternatively we could just provide a command for each custom component instead of the repo and componentPath.

When using this feature in in combination with the git-sync feature that allows for your Home Assistant configuration
to be pulled from a git repo, you must create a symlink from custom_components to ../custom_components in your
Home Assistant configuration git repository. This patch will create a /custom_components directory in the Pod such
that the symlink will be satisfied. All custom components will be cloned into the /custom_componenets directory.

Should customcomponents be nested under git? This way we can use the common image/tag for the init container image.

This patch was tested with the following configuration and custom components.

At the time of this writing, the repositories are cloned, but I have done no functional testing.

customcomponents:
  enabled: true

  ## we just use the hass-configurator container image
  ## you can use any image which has git and openssh installed
  ##
  image:
    repository: causticlab/hass-configurator-docker
    tag: 0.3.5-x86_64
    pullPolicy: IfNotPresent

  syncPath: /custom_components

  components:
  - name: ewelink
    repo: "https://github.com/peterbuga/HASS-sonoff-ewelink"
    componentPath: HASS-sonoff-ewelink/sonoff
  - name: wyzesense
    repo: "https://github.com/kevinvincent/ha-wyzesense"
    componentPath: ha-wyzesense/custom_components/wyzesense
  - name: hacs
    command: ['sh', '-c', 'cd /config ; echo "pwd" ; pwd ; /usr/bin/curl -sfSL https://hacs.xyz/install | bash - ; echo "ls ." ; ls -la . ; echo "ls /config/custom_components" ; ls -la /config/custom_components']