Allow helm chart to pull secrets.yaml from (private) git repo
This patch grants the user the ability to provide a private git repository
that contains their secrets.yaml. I could not find a way to do this with my
understanding of the current functionality of this helm chart. The problem is
that a user may want to share their Home Assistant configuration in a public
git repository but keep their secrets.yaml in a private git repository.

This patch provides a solution by pulling a repository into the /secrets directory
within the Home Assistant Pod. When used with the git-sync feature, the user
should create a symlink from secrets.yaml to ../secrets/secrets.yaml in their
Home Assistant configuration git repository. After the init container pulls the
secrets repository the symlink is satisfied.

Is there a better way to do this? This was the solution I came up with. It's similar
to how I'm doing my custom components patch. If there is a better way, I'd
like to know about it.